app/requirements.txt
paddleocr==2.7.0.3
paddlepaddle==2.4.2 ; sys_platform == "win32"
sqlalchemy>=1.4
psycopg2-binary
pdf2image
Pillow 